requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
requests
sentence-transformers
faiss-cpu
python-jose[cryptography]
python-multipart
jinja2
passlib[bcrypt]
PyPDF2
tinydb